`long` is a better idea

+++ src/Formatter/CoopDuration.cs
@@ -2,9 +2,9 @@ namespace WHAL_Int.Formatter;
 
 public class CoopDuration : IComparable<CoopDuration>
 {
-    public int DurationInSeconds { get; set; } = 0;
+    public long DurationInSeconds { get; set; } = 0;
 
-    public CoopDuration(int durationInSeconds) => DurationInSeconds = durationInSeconds;
+    public CoopDuration(long durationInSeconds) => DurationInSeconds = durationInSeconds;
 
     public string Format()
     {
